requirements.txt
dotenv
livekit-api
livekit-agents
livekit-plugins-bey
livekit-plugins-google
requests
aiohttp
fastapi
uvicorn